message, user, channel 중복/null/blank 체크

+++ src/main/java/com/sprint/mission/discodeit/service/file/FileMessageService.java
@@ -15,6 +15,12 @@ public class FileMessageService implements MessageService {
 
     @Override
     public UUID create(Message message) {
+        if(message == null){
+            throw new IllegalArgumentException("메세지가 null입니다.");
+        }
+        if(message.getContent() ==null || message.getContent().isBlank()){
+            throw new IllegalArgumentException("메세지 내용이 null이거나 blank입니다.");
+        }
         return messageRepository.create(message);
     }
 
@@ -32,12 +38,25 @@ public class FileMessageService implements MessageService {
     @Override
     public void update(UUID id, String messageContent) {
         Message message = messageRepository.read(id);
+        if(message == null){
+            throw new IllegalArgumentException("존재하지 않는 메세지입니다.");
+        }
+        if(messageContent == null){
+            throw new IllegalArgumentException("메세지 내용이 null입니다.");
+        }
+        if(messageContent.isBlank()){
+            throw new IllegalArgumentException("메세지 내용이 blank입니다.");
+        }
         message.updateContent(messageContent);
         messageRepository.create(message);
     }
 
     @Override
     public void delete(UUID id) {
+        Message message = messageRepository.read(id);
+        if(message == null){
+            throw new IllegalArgumentException("존재하지 않는 메세지입니다.");
+        }
         messageRepository.delete(id);
     }
 }
